Update BoardSquare and DisplayPiece Code
The BoardSquare Class Definition is Updated to also include a certain ChessPiece.
All the Chess Pieces are Assigned to Default Stating Position Board Squares.

The PieceDisplay is Updated to Display a ChessPiece's First Letter, (Uppercase for White, Lowercase for Black), where they are Positioned.
The Enumeration Values are Needed to be Converted to String, from where the First Letter is Extracted.
The Previous Manual Setup is Removed, And Now its Dynamic.

The Remaining Central Empty Squares are Initialized with a temporary ChessPiece Object (Initilialization was Necessary).