Complete Code Simplification
The DisplayPiece Function's Code is Now Great.
It Takes the Square's Rank & File, where the Piece would be Displayed.
It Returns a Character as a Piece's Image/Symbol on CLI-Terminal.

The ChessPiece Definition now includes a PieceName (String), before it were only Integer.
The PieceToString Function replaced by getName classFunction. (No Need for setName Function)

The PrintBoard Function was already Working Fine (But Code-Repeat Ineffeciency is still there).
The Type String is Replaced by Enumerated Color (Now, No Chance of Wrong Input Error).

The BoardSquare Class Definition now includes isEmpty (Boolean), to check if its Empty or Occupied by a Chess-Piece.

The BoardSquares and ChessPieces are Initialized in a initBoard Function (rather than in Main)
Instead of a Single Piece Array of 32-index; Now made Separate Piece-Object's Arrays.
(Because the Single Array of All Pieces wasn't making A Good Index-Sequence.)

Also, the Enum Values are Set to Uppercase, for Differentiation.