Game Sequence & Player Input Code
The Complete Game-Sequence is defined, in the Game Function.
The Player can now Input Move, in the turn Function.
The Input is Filtered:
     > To remain In the Board-Squares.
     > To Move only Own Piece (not of Opponent's or an Empty Square)
     > To Not Land on Own's Piece

After Input Filtering, the Piece is Ready to Move.
(But the Actual Piece Movement didn't Happened yet)

Both the Players would Continue Inputing their Turns, until the Game is Over.
The gameover Algorithm is also not Defined yet.