Class-Definition and Turn-Input Update
The ChessPiece setID Function now also Derive the PieceName
(previously the PieceName was only Converting at Runtime by return Function, which means we could not use the PieceName unless we Get it Somewhere First)
such as string n = getName(); or simply, getName();

The BoardSquare file is now a Charcter, the Integer One is renamed to fileNumber.
The setFileNumber also derive the file-Character.

The Player's Turn now Input the Files as Character (Also could be Input With or Without Spaces).
such a2a3 to Move a2 Piece to a3 Square

This Move Turn Input could also be Implemented without Updating Class Definition.
But Its Good to have FileCharacter, so it becomes an Intrinsic Property of a BoardSquare.