ChessPieces Movement Algorithm
The Player-Input's Source and Destined Squares are Now in Global Scope, to be Accessed by MovePiece Functions.
The Chess-Piece Class is Updated to include a Move Function. From where it handles the Movement of Other Pieces.

The Pawns can Move Straight-Step, Capture with Diagonal Step, and Move Double-Step in First-Move
But they can also Jump Other Pieces in Double-Step, for Now.
The pawnMoved Boolean is Defined in the Class-Definition, just to check Pawn's First-Move.

The KnightMove & KingMove Movement-Algorithm is Great.
The Knight & King can Move like they should Move.

The Bishop, Rook, and Queen can Move like usual, but they can also Jump Other Pieces along the Way, for Now.