Apply Class Inheritance
The Inheritance is Defined in the Classes. The BoardSquare is the Base.
The initBoard is completety redefined to have the ChessPiece their own Rank & File (Rather than Assigning Each ChessPiece on a certain Square).
The BoardSquare Class is Only to tell apart the Empty Squares. (Even this might be Removed in the Future Updates)

An Algorithm (getPiece Function) is Defined to Get the ChessPiece on the given Rank/File.
The Player Input and Move Variables are Simplified.