GameOver & PlayAgain Algorithm
ChesspIece Captured is Defined, to check if the Piece is Captured.

The Chesspiece Constrcutor is also Defined.
The gameOver Boolean-Variable is Defined to check if the Game is Over.
The Game would Over, if the KIng gets Captured.
The ChessPiece Movement won't be Bound, if the King is in Check. as there's No COncept of King-Check for Now.

The PlayAgain Algorithm is also Implemented.